387: Reuse buildAuthorUrl util in author entry utils. Move buildAuthorUrl and removeAuthorsSuffixes to article utils file.
