+++ aemedge/scripts/utils.js
@@ -1,5 +1,4 @@
 import { div } from './dom-builder.js';
-import { toClassName } from './aem.js';
 
 function formatDate(inputDate) {
   const options = { year: 'numeric', month: 'long', day: 'numeric' };
@@ -16,32 +15,4 @@ function containerize(container, targetClass) {
   }
 }
 
-const defaultSuffixes = ['PhD', 'Ph.D.'];
-function removeAuthorsSuffixes(authors, suffixes = defaultSuffixes) {
-  if (!authors) {
-    return '';
-  }
-  let authorsWithoutSuffixes = authors;
-  suffixes.forEach((suffix) => {
-    /**
-     * Jane Doe, PhD,
-     * Jane Doe, PhD{eol}
-     * Jane Doe PhD,
-     * Jane Doe PhD{eol}
-     * Jane Doe, Ph.D.,
-     * Jane Doe, Ph.D.{eol}
-     * Jane Doe Ph.D.,
-     * Jane Doe Ph.D.{eol}
-     */
-    authorsWithoutSuffixes = authorsWithoutSuffixes.replaceAll(new RegExp(`,*\\s*${suffix}(?=,|$)`, 'g'), '');
-  });
-  return authorsWithoutSuffixes;
-}
-
-function buildAuthorUrl(author) {
-  return `/author/${toClassName(author.trim()).replaceAll('-', '')}`;
-}
-
-export {
-  formatDate, containerize, removeAuthorsSuffixes, buildAuthorUrl,
-};
+export { formatDate, containerize };
